requirements.txt
discord.py>=2.3.2
python-dotenv==1.0.0
aiohttp==3.9.5
psutil==5.9.8
yt-dlp>=2023.3.4
PyNaCl>=1.5.0
spotipy>=2.23.0
